HSM 3.20.3 (#389)
* Small collector refactiong

* Some fixes have been added

* ToEmpty extension has been removed. Status comment has been fixed

* Status condition (#384)

* Added condition constructor for sensor status

* Added AlertProperty enum

---------


* Property compare and GetLastValue has been fixed

* ReleaseNote has been uploaded

* Time to live (#390)

* Added display mode for ttl policy in sensor

* Some fixes and refactoring

* Fixed custom period for time intervals with predefined lists

* TTL Policy update (#387)

* Refactoring

* TTLPolicy update has been added. Default TTL icon has been changed

* Save TTLPolicyEntity logic has been added

* Fixed attach sensor policies

* BuildState method has been added for Default policy

* PredefinedIntervals lists are changed to HashSet

* Added display comment for sensor ttl

* Fixed ttl interval in constructo (meta info)

* Fixed UI for product/node/sensor ttl

* Added enable/disable Add TTL link logic

* Added logic for updating sensor ttl

* Added logic for saving node/product ttl

* Added TTL for folder

* TTL update has been fixed (#391)

* TTL update has been fixed

* .

* Some fixes have been added

* Build state with default condition fix has been added

---------


* Fixed remark

---------


* UI fixes (#388)

* TAM-969. HSM, Tree: 9+ icons counter should be changed to infinity sign

* Increased version of server

* TAM-967. HSM, Tree: Alert icons should be hidden for folders

* TAM-968. HSM, Grid, List: Alert icons should be added

---------


* TAM-962, TAM-950  (#392)

* Min status level removed from View.

* TAM-950 HSM, Folder: Grid should be fixed

---------


---------

+++ src/server/HSMServer/Model/TreeViewModels/SensorNodeViewModel.cs
@@ -66,10 +66,9 @@ namespace HSMServer.Model.TreeViewModel
             FileNameString = GetFileNameString(model.Type, ShortStringValue);
 
             if (model is DoubleSensorModel or IntegerSensorModel or DoubleBarSensorModel or IntegerBarSensorModel)
-                DataAlerts[Type] = model.Policies.Select(p => BuildAlert(p, model)).ToList();
+                DataAlerts[(byte)Type] = model.Policies.Select(p => BuildAlert(p, model)).ToList();
 
             AlertIcons.Clear();
-
             foreach (var alert in model.PolicyResult)
             {
                 var icon = alert.Icon;
